requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
firebase-admin==7.1.0
pandas==2.1.4
schedule==1.2.2
apscheduler==3.11.0
python-dotenv==1.1.0
gunicorn==21.2.0
lxml==4.9.3
html5lib==1.1
nltk==3.8.1
numpy==1.24.4
mysql-connector-python==8.1.0
